Remove dependency on 32-bit integers for underlying buffer

+++ include/crcham/codeword.hpp
@@ -7,13 +7,13 @@
 namespace crcham {
 
 __device__ __host__
-void permute(uint32_t* arr, size_t len, uint64_t n, size_t m, size_t k);
+void permute(uint8_t* arr, size_t len, uint64_t n, size_t m, size_t k);
 
 __device__ __host__
 uint64_t extract(uint8_t* arr, size_t len, size_t bits, size_t polybits);
 
 __device__ __host__
-size_t popcount(const uint32_t* arr, size_t len);
+size_t popcount(const uint8_t* arr, size_t len);
 
 }
